+++ src/client/components/saaCreate/saaCreate.js
@@ -3,7 +3,6 @@
  */
 
 import { pwixI18n } from 'meteor/pwix:i18n';
-import { ReactiveVar } from 'meteor/reactive-var';
 
 import 'meteor/pwix:accounts-ui';